+++ Controllers/BookingController.cs
@@ -1,9 +1,15 @@
 ﻿using API.DTOs.Bookings;
-using BookingManagementApp.Contracts;
-using BookingManagementApp.Models;
+using API.DTOs.Universities;
+using API.Contracts;
+using API.Models;
+using API.Repositories;
 using Microsoft.AspNetCore.Mvc;
+using API.DTOs.Accounts;
+using API.Utilities.Handlers;
+using API.Utilities.Handlers.Exceptions;
+using API.DTOs.AccountRoles;
 
-namespace BookingManagementApp.Controllers
+namespace API.Controllers
 {
     //membuat endpoint routing untuk booking controller 
     [ApiController]
@@ -24,11 +30,11 @@ namespace BookingManagementApp.Controllers
             var result = _bookingRepository.GetAll();
             if (!result.Any())
             {
-                return NotFound("Data Not Found");
+                return NotFound(new ResponseNotFoundHandler("Data Not Found"));
             }
             var data = result.Select(i => (BookingDto)i);
 
-            return Ok(data);
+            return Ok(new ResponseOkHandler<IEnumerable<BookingDto>>(data));
         }
         //method get dari http untuk getByGuid booking
         [HttpGet("{guid}")]
@@ -37,47 +43,73 @@ namespace BookingManagementApp.Controllers
             var result = _bookingRepository.GetByGuid(guid);
             if (result is null)
             {
-                return NotFound("Data Not Found");
+                return NotFound(new ResponseNotFoundHandler("Data Not Found"));
 
             }
-            return Ok((BookingDto) result);
+            return Ok(new ResponseOkHandler<BookingDto>((BookingDto)result));
         }
         //method post dari http untuk create booking
         [HttpPost]
-        public IActionResult Create(CreateBookingDto booking)
+        public IActionResult Create(CreateBookingDto createBookingDto)
         {
-            var result = _bookingRepository.Create(booking);
-            if (result is null)
+            try
             {
-                return BadRequest("Failed To Create Data");
+                Booking toCreate = createBookingDto;
+                var result = _bookingRepository.Create(toCreate);
+                return Ok(new ResponseOkHandler<BookingDto>((BookingDto)result));
+
             }
-            return Ok((BookingDto) result);
+            catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
+            }
         }
 
         //method put dari http untuk Update booking
         [HttpPut]
-        public IActionResult Update(BookingDto booking)
+        public IActionResult Update(BookingDto bookingDto)
         {
-            var result = _bookingRepository.Update(booking);
-            if (!result)
+            try
             {
-                return BadRequest("Failed To Update Data");
+                var entity = _bookingRepository.GetByGuid(bookingDto.Guid);
+                if (entity is null)
+                {
+                    return NotFound(new ResponseNotFoundHandler("Data Not Found"));
+
+                }
+                Booking toUpdate = bookingDto;
+                toUpdate.CreatedDate = entity.CreatedDate;
+                var result = _bookingRepository.Update(bookingDto);
+                return Ok(new ResponseOkHandler<String>("Data Updated"));
+
+            }
+            catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
             }
 
-            return Ok(result);
         }
         //method delete dari http untuk delete booking
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-            var booking = _bookingRepository.GetByGuid(guid);
-            var result = _bookingRepository.Delete(booking);
-            if (!result)
+            try
             {
-                return BadRequest("Failed To Delete Data");
+                var booking = _bookingRepository.GetByGuid(guid);
+                if (booking is null)
+                {
+                    return NotFound(new ResponseNotFoundHandler("Data Not Found"));
+
+                }
+                var result = _bookingRepository.Delete(booking);
+                return Ok(new ResponseOkHandler<String>("Data Deleted"));
+
+            }
+            catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
             }
 
-            return Ok(result);
         }
     }
 }
