+++ Controllers/EducationController.cs
@@ -1,8 +1,14 @@
-﻿using BookingManagementApp.Contracts;
-using BookingManagementApp.Models;
+﻿using API.DTOs.Educations;
+using API.DTOs.Universities;
+using API.Contracts;
+using API.Models;
+using API.Repositories;
 using Microsoft.AspNetCore.Mvc;
+using API.Utilities.Handlers;
+using API.Utilities.Handlers.Exceptions;
+using API.DTOs.AccountRoles;
 
-namespace BookingManagementApp.Controllers
+namespace API.Controllers
 {
     //membuat endpoint routing untuk education controller 
     [ApiController]
@@ -23,10 +29,11 @@ namespace BookingManagementApp.Controllers
             var result = _educationRepository.GetAll();
             if (!result.Any())
             {
-                return NotFound("Data Not Found");
+                return  NotFound(new ResponseNotFoundHandler("Data NotFound"));
             }
+            var data = result.Select(i => (EducationDto)i);
 
-            return Ok(result);
+            return Ok(new ResponseOkHandler<IEnumerable<EducationDto>>(data));
         }
         //method get dari http untuk getByGuid education
         [HttpGet("{guid}")]
@@ -35,47 +42,71 @@ namespace BookingManagementApp.Controllers
             var result = _educationRepository.GetByGuid(guid);
             if (result is null)
             {
-                return NotFound("Data Not Found");
+                return NotFound(new ResponseNotFoundHandler("Data NotFound"));
 
             }
-            return Ok(result);
+            return Ok(new ResponseOkHandler<EducationDto>((EducationDto) result));
         }
         //method post dari http untuk create education
         [HttpPost]
-        public IActionResult Create(Education education)
+        public IActionResult Create(CreateEducationDto createEducationDto)
         {
-            var result = _educationRepository.Create(education);
-            if (result is null)
+            try
             {
-                return BadRequest("Failed To Create Data");
+                Education toCreate = createEducationDto;
+                var result = _educationRepository.Create(toCreate);
+
+                return Ok(new ResponseOkHandler<EducationDto>((EducationDto)result));
             }
-            return Ok(result);
+            catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
+            }
+                
+        
         }
 
         //method put dari http untuk Update education
         [HttpPut]
-        public IActionResult Update(Education education)
+        public IActionResult Update(EducationDto educationDto)
         {
-            var result = _educationRepository.Update(education);
-            if (!result)
+            try
             {
-                return BadRequest("Failed To Update Data");
-            }
+                var entity = _educationRepository.GetByGuid(educationDto.Guid);
+                if (entity is null)
+                {
+                    return NotFound(new ResponseNotFoundHandler("Data Not Found"));
 
-            return Ok(result);
+                }
+                Education toUpdate = educationDto;
+                toUpdate.CreatedDate = entity.CreatedDate;
+                var result = _educationRepository.Update(educationDto);
+
+                return Ok(new ResponseOkHandler<String>("Data Updated"));
+
+            }  catch (Exception e) {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
+            }
         }
         //method delete dari http untuk delete education
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-            var education = _educationRepository.GetByGuid(guid);
-            var result = _educationRepository.Delete(education);
-            if (!result)
-            {
-                return BadRequest("Failed To Delete Data");
-            }
+            try
+            { 
+                var education = _educationRepository.GetByGuid(guid);
+                if (education is null)
+                {
+                    return NotFound(new ResponseNotFoundHandler("Data Not Found"));
 
-            return Ok(result);
+                }
+                var result = _educationRepository.Delete(education);
+                return Ok(new ResponseOkHandler<String>("Data Deleted"));
+
+            } catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
+            }
         }
     }
 }
