+++ Controllers/EmployeeController.cs
@@ -1,8 +1,17 @@
-﻿using BookingManagementApp.Contracts;
-using BookingManagementApp.Models;
+﻿using API.DTOs.Educations;
+using API.DTOs.Employees;
+using API.DTOs.Universities;
+using API.Contracts;
+using API.Models;
+using API.Repositories;
 using Microsoft.AspNetCore.Mvc;
+using API.Utilities.Handlers;
+using API.DTOs.Accounts;
+using API.DTOs.Rooms;
+using API.Utilities.Handlers.Exceptions;
+using API.DTOs.AccountRoles;
 
-namespace BookingManagementApp.Controllers
+namespace API.Controllers
 {
     //membuat endpoint routing untuk employee controller 
     [ApiController]
@@ -23,10 +32,10 @@ namespace BookingManagementApp.Controllers
             var result = _employeeRepository.GetAll();
             if (!result.Any())
             {
-                return NotFound("Data Not Found");
+                return NotFound(new ResponseNotFoundHandler("Data Not Found"));
             }
-
-            return Ok(result);
+            var data = result.Select(i => (EmployeeDto)i);
+            return Ok(new ResponseOkHandler<IEnumerable<EmployeeDto>>(data));
         }
         //method get dari http untuk getByGuid employee
         [HttpGet("{guid}")]
@@ -35,47 +44,70 @@ namespace BookingManagementApp.Controllers
             var result = _employeeRepository.GetByGuid(guid);
             if (result is null)
             {
-                return NotFound("Data Not Found");
+                return NotFound(new ResponseNotFoundHandler("Data Not Found"));
 
             }
-            return Ok(result);
+            return Ok(new ResponseOkHandler<EmployeeDto>((EmployeeDto)result));
         }
         //method post dari http untuk create employee
         [HttpPost]
-        public IActionResult Create(Employee employee)
+        public IActionResult Create(CreateEmployeeDto createEmployeeDto)
         {
-            var result = _employeeRepository.Create(employee);
-            if (result is null)
+            try
             {
-                return BadRequest("Failed To Create Data");
+                Employee toCreate = createEmployeeDto;
+                toCreate.NIK = GenerateNIKHandler.GenerateNIK(_employeeRepository);
+                var result = _employeeRepository.Create(toCreate);
+                return Ok(new ResponseOkHandler<EmployeeDto>((EmployeeDto)result));
+
             }
-            return Ok(result);
+            catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
+            }
         }
 
         //method put dari http untuk Update employee
         [HttpPut]
-        public IActionResult Update(Employee employee)
+        public IActionResult Update(EmployeeDto employeeDto)
         {
-            var result = _employeeRepository.Update(employee);
-            if (!result)
+            try
             {
-                return BadRequest("Failed To Update Data");
+                var entity = _employeeRepository.GetByGuid(employeeDto.Guid);
+                if (entity is null)
+                {
+                    return NotFound(new ResponseNotFoundHandler("Data Not Found"));
+
+                }
+                Employee toUpdate = employeeDto;
+                toUpdate.CreatedDate = entity.CreatedDate;
+                var result = _employeeRepository.Update(employeeDto);
+                return Ok(new ResponseOkHandler<String>("Data Updated"));
+
+            }
+            catch (Exception e)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
             }
 
-            return Ok(result);
         }
         //method delete dari http untuk delete employee
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
-        {
-            var employee = _employeeRepository.GetByGuid(guid);
-            var result = _employeeRepository.Delete(employee);
-            if (!result)
+        { try { 
+                var employee = _employeeRepository.GetByGuid(guid);
+                if (employee is null)
+                {
+                    return NotFound(new ResponseNotFoundHandler("Data Not Found"));
+
+                }
+                var result = _employeeRepository.Delete(employee);
+                return Ok(new ResponseOkHandler<String>("Data Deleted"));
+        }catch (Exception e)
             {
-                return BadRequest("Failed To Delete Data");
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseInternalServerErrorHandler("Failed to Create Data", e.Message));
             }
 
-            return Ok(result);
         }
     }
 }
